Add LiquidInfrastructureMulticlaim feature
The ERC20 now accepts a multiclaim address in the constructor to
whitelist a single contract for use with the claimRevenueFor() function.
With this addition, users may call the
Multiclaim.claimRevenueMulti() function with all of the ERC20s they want
to claim revenue on. This batches the calls into one transaction.